move attendance files
