+++ apps/dashboard/src/app/(dashboard)/event/[id]/attendees-page.tsx
@@ -1,10 +1,10 @@
 import type { Attendance } from "@dotkomonline/types"
 import { Box, Divider, Title } from "@mantine/core"
 import type { FC } from "react"
-import QrCodeScanner from "src/modules/attendance/components/attendance-page/QrCodeScanner"
 import { UserSearch } from "../../../../components/molecules/UserSearch/UserSearch"
-import { openCreateManualUserAttendModal } from "../../../../modules/attendance/modals/manual-user-attend-modal"
 import { AllAttendeesTable } from "../components/all-users-table"
+import { openCreateManualUserAttendModal } from "../components/manual-user-attend-modal"
+import QrCodeScanner from "../components/qr-code-scanner"
 import { useEventDetailsContext } from "./provider"
 
 export const AttendeesPage: FC = () => {
