+++ apps/dashboard/src/app/(dashboard)/event/components/InfoBox.tsx
