+++ apps/dashboard/src/app/(dashboard)/event/components/attendance-registered-modal.tsx
