+++ apps/dashboard/src/app/(dashboard)/event/components/error-attendance-registered-modal.tsx
