+++ apps/dashboard/src/app/(dashboard)/event/components/pools-box.tsx
@@ -1,9 +1,9 @@
 import type { AttendancePool } from "@dotkomonline/types"
 import { Box, Button, Card, Flex, Text } from "@mantine/core"
 import type { FC } from "react"
-import { notifyFail } from "../../../../app/notifications"
-import { openEditPoolModal } from "../../modals/edit-pool-modal"
-import { useDeletePoolMutation } from "../../mutations/use-pool-mutations"
+import { notifyFail } from "../../../notifications"
+import { useDeletePoolMutation } from "../mutations"
+import { openEditPoolModal } from "./edit-pool-modal"
 
 interface PoolsBoxProps {
   pools: AttendancePool[]
