+++ apps/dashboard/src/app/(dashboard)/event/components/utils.ts
