+++ apps/dashboard/src/app/(dashboard)/event/mutations.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query"
 import { useRouter } from "next/navigation"
 import { useTRPC } from "../../../trpc"
-import { useQueryNotification } from "../../notifications"
+import { useQueryGenericMutationNotification, useQueryNotification } from "../../notifications"
 
 export const useAddCompanyToEventMutation = () => {
   const trpc = useTRPC()
@@ -99,6 +99,194 @@ export const useRemoveCompanyFromEventMutation = () => {
           message: "Bedriften har blitt fjernet fra arrangørlisten.",
         })
       },
+    })
+  )
+}
+
+export const useDeletePoolMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "delete",
+  })
+
+  return useMutation(
+    trpc.event.attendance.deletePool.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useCreatePoolMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "create",
+  })
+
+  return useMutation(
+    trpc.event.attendance.createPool.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useUpdatePoolMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "update",
+  })
+
+  return useMutation(
+    trpc.event.attendance.updatePool.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useRegisterForEventMutation = () => {
+  const trpc = useTRPC()
+  const notification = useQueryNotification()
+
+  return useMutation(
+    trpc.event.attendance.registerForEvent.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Melder på bruker",
+          message: "Brukeren blir meldt på arrangementet.",
+        })
+      },
+      onSuccess: (data) => {
+        notification.complete({
+          title: "Påmelding vellykket",
+          message: "Bruker ble påmeldt arrangementet.",
+        })
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under påmelding: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
+export const useDeregisterForEventMutation = () => {
+  const trpc = useTRPC()
+  const notification = useQueryNotification()
+
+  return useMutation(
+    trpc.event.attendance.adminDeregisterForEvent.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Melder av bruker",
+          message: "Brukeren blir meldt av arrangementet.",
+        })
+      },
+      onSuccess: () => {
+        notification.complete({
+          title: "Avmelding vellykket",
+          message: "Bruker ble meldt av arrangementet.",
+        })
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under avmelding: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
+export const useUpdateEventAttendanceMutation = () => {
+  const trpc = useTRPC()
+  const notification = useQueryNotification()
+  return useMutation(
+    trpc.event.attendance.registerAttendance.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Oppdaterer oppmøte...",
+          message: "Brukerens oppmøte blir oppdatert.",
+        })
+      },
+      onSuccess: (data) => {
+        notification.complete({
+          title: "Oppmøte oppdatert",
+          message: `Oppmøte er ${data.attended ? "registrert" : "fjernet"}. `,
+        })
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under oppdatering av oppmøte: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
+export const useAddAttendanceMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "create",
+  })
+
+  return useMutation(
+    trpc.event.addAttendance.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useUpdateAttendanceMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "update",
+  })
+
+  return useMutation(
+    trpc.event.attendance.updateAttendance.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useMergeAttendanceMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "update",
+  })
+
+  return useMutation(
+    trpc.event.attendance.mergeAttendancePools.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
+    })
+  )
+}
+
+export const useUpdateSelectionResponsesMutation = () => {
+  const trpc = useTRPC()
+  const { fail, loading, complete } = useQueryGenericMutationNotification({
+    method: "update",
+  })
+
+  return useMutation(
+    trpc.event.attendance.updateSelectionResponses.mutationOptions({
+      onError: fail,
+      onMutate: loading,
+      onSuccess: complete,
     })
   )
 }
