--- apps/dashboard/src/modules/attendance/mutations/use-attendance-mutations.ts
@@ -1,64 +1,0 @@
-import { useQueryGenericMutationNotification } from "../../../app/notifications"
-import { useTRPC } from "../../../trpc"
-
-import { useMutation } from "@tanstack/react-query"
-
-export const useAddAttendanceMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "create",
-  })
-
-  return useMutation(
-    trpc.event.addAttendance.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
-
-export const useUpdateAttendanceMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "update",
-  })
-
-  return useMutation(
-    trpc.event.attendance.updateAttendance.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
-
-export const useMergeAttendanceMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "update",
-  })
-
-  return useMutation(
-    trpc.event.attendance.mergeAttendancePools.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
-
-export const useUpdateSelectionResponsesMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "update",
-  })
-
-  return useMutation(
-    trpc.event.attendance.updateSelectionResponses.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
