--- apps/dashboard/src/modules/attendance/mutations/use-pool-mutations.ts
@@ -1,49 +1,0 @@
-import { useQueryGenericMutationNotification } from "../../../app/notifications"
-import { useTRPC } from "../../../trpc"
-
-import { useMutation } from "@tanstack/react-query"
-
-export const useDeletePoolMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "delete",
-  })
-
-  return useMutation(
-    trpc.event.attendance.deletePool.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
-
-export const useCreatePoolMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "create",
-  })
-
-  return useMutation(
-    trpc.event.attendance.createPool.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
-
-export const useUpdatePoolMutation = () => {
-  const trpc = useTRPC()
-  const { fail, loading, complete } = useQueryGenericMutationNotification({
-    method: "update",
-  })
-
-  return useMutation(
-    trpc.event.attendance.updatePool.mutationOptions({
-      onError: fail,
-      onMutate: loading,
-      onSuccess: complete,
-    })
-  )
-}
